Rough Roadmap Creation
Adding is a rough roadmap to complete this project. Please help by adding in anything you think may be necessary to complete this project. I believe there may be a way to use the NFTs themselves as a way to store the database info but I haven't made it that far in my NFT travels yet. If you know how and which to contribute please join us!