Working Upvote/Downvote
Upvote/Downvote system records rating_action as either "updoot" or "boop", rating_date as now, post_id as the post that was voted on, user_id as the logged in user in table rating_info. This is to ensure only one rating_action  can be recorded for each specific post_id. The user can update their rating_action as long as the post_id is active.